check if tekton is installed
